# .stryker-tmp/sandbox-bCteHS/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: VibeX API Contract
  description: |
    VibeX 全栈项目 API 契约文档
    
    ## 版本历史
    - v1.1.0 (2026-03-28): 从 14 端点扩展至 90+ 端点，覆盖全部前端实际调用
    - v1.0.0 (初始): 基础 14 端点定义
    
    ## 路径约定
    - 所有路径相对于 `/api` 前缀（如 `/auth/login` → `/api/auth/login`）
    - 服务器基础路径: `/api`
    - 统一无 v1 前缀（v1 双写路由已废弃，标注 `deprecated: true`）
    
    ## 路由分类
    - **前端调用**: 前端 services/api/ 模块中实际调用
    - **后端独有**: `x-backend-only: true`，前端未调用但后端已实现
    - **废弃路由**: `deprecated: true`，v1 双写兼容层
    
    ## 16 个 Tag 分组
    | Tag | 前缀 | 说明 |
    |-----|------|------|
    | Auth | /auth | 认证 |
    | Users | /users | 用户 |
    | Projects | /projects | 项目管理 |
    | Requirements | /requirements | 需求 |
    | Clarifications | /clarifications | 澄清问答 |
    | DomainEntities | /domain-entities, /domains | 域实体 |
    | EntityRelations | /entity-relations | 实体关系 |
    | DDD | /ddd | DDD 限界上下文/域模型/业务流程 |
    | Design | /clarify, /domain, /flow, /pages, /prototype, /design | 设计流程 |
    | Flows | /flows | 流程图 |
    | PrototypeSnapshots | /prototype-snapshots | 原型快照 |
    | Agents | /agents | Agent |
    | Pages | /pages | 页面 |
    | Chat | /chat | AI 对话 |
    | Plan | /plan | 计划分析 |
    | Diagnosis | /diagnosis | 需求诊断 |
    | BackendOnly | 各类后端独有路由 | 未被前端调用的后端路由 |
    | Deprecated | /api/v1/* | v1 兼容层（废弃） |
  version: 1.1.0
  contact:
    name: VibeX Team
servers:
  - url: /api
    description: API 基础路径
...
paths:
...
  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: 获取单个项目
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Projects]
      summary: 更新项目
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Projects]
      summary: 删除项目（永久删除）
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /messages:
    get:
      tags: [Messages]
      summary: 获取消息列表
      operationId: getMessages
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Messages]
      summary: 创建消息
      operationId: createMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
  /messages/{messageId}:
    delete:
      tags: [Messages]
      summary: 删除消息
      operationId: deleteMessage
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # TAG: CHAT
  # =============================================================================
...
  /v1/projects/{projectId}:
    get:
      tags: [Deprecated]
      summary: 获取单个项目（v1 兼容，已废弃）
      deprecated: true
      operationId: getProjectV1
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
    put:
      tags: [Deprecated]
      summary: 更新项目（v1 兼容，已废弃）
      deprecated: true
      operationId: updateProjectV1
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: 更新成功（兼容模式）
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
    delete:
      tags: [Deprecated]
      summary: 删除项目（v1 兼容，已废弃）
      deprecated: true
      operationId: deleteProjectV1
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 删除成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /v1/messages:
    get:
      tags: [Deprecated]
      summary: 获取消息列表（v1 兼容，已废弃）
      deprecated: true
      operationId: getMessagesV1
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功（兼容模式）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      tags: [Deprecated]
      summary: 创建消息（v1 兼容，已废弃）
      deprecated: true
      operationId: createMessageV1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '201':
          description: 创建成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /v1/messages/{messageId}:
    delete:
      tags: [Deprecated]
      summary: 删除消息（v1 兼容，已废弃）
      deprecated: true
      operationId: deleteMessageV1
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功（兼容模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: 项目 ID
  schemas:
...
    Project:
      type: object
      required:
        - id
        - name
        - userId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        userId:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        pages:
          type: array
          items:
            $ref: '#/components/schemas/Page'
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...
    Page:
      type: object
      required:
        - id
        - name
        - projectId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        content:
          type: string
          nullable: true
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        project:
          $ref: '#/components/schemas/Project'
...
    Message:
      type: object
      required:
        - id
        - role
        - content
        - projectId
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        projectId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    MessageCreate:
      type: object
      required:
        - content
        - projectId
      properties:
        content:
          type: string
        projectId:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
          default: user

    # ===== 计划分析相关 =====
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...