# vibex-backend/src/app/api/v1/messages/[messageId]/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { prisma } from '@/lib/prisma';
import { getAuthUserFromRequest } from '@/lib/authFromGateway';
import { getEnv } from '@/lib/env';

import { safeError } from '@/lib/log-sanitizer';


export async function DELETE(
  request: NextRequest,
  { params }: { params: Promise<{ messageId: string }> }
) {
  try {
    const env = getEnv();
    const auth = getAuthUserFromRequest(request, env.JWT_SECRET);
    if (!auth) {
      return NextResponse.json(
        { success: false, error: 'Not authenticated', code: 'UNAUTHORIZED' },
        { status: 401 }
      );
    }

    const { messageId } = await params;

    // Find the message and verify ownership through project
    const message = await prisma.message.findUnique({
      where: { id: messageId },
      include: {
        project: true,
      },
    });

    if (!message) {
      return NextResponse.json(
        { success: false, error: 'Message not found', code: 'NOT_FOUND' },
        { status: 404 }
      );
    }

    // Only allow users to delete their own messages
    if (message.project.userId !== auth.userId) {
      return NextResponse.json(
        { success: false, error: 'Forbidden', code: 'FORBIDDEN' },
        { status: 403 }
      );
    }

    await prisma.message.delete({
      where: { id: messageId },
    });

    return NextResponse.json({
      success: true,
      data: { message: 'Message deleted successfully' },
    });
  } catch (error) {
    safeError('Delete message error:', error);
    return NextResponse.json(
      { success: false, error: 'Internal server error', code: 'INTERNAL_ERROR' },
      { status: 500 }
    );
  }
}